refactor: remove node lock implementation
- Deleted the entire node lock implementation from `node_lock.rs`, which included mutex handling for process crash monitoring across platforms (Windows, Linux, macOS).
- Removed associated functions for acquiring, checking, and releasing locks, as well as the `NodeLockWatcher` struct and its methods.
- Cleaned up related tests that depended on the node lock functionality.

fix: improve error handling in Node.js dispatch

- Updated the `call` function in `nodejs_dispatch.rs` to return an `Err` instead of panicking when the dispatch function is not set, improving robustness.

docs: add comments to ReconnectManager

- Added documentation to the `new` method in `ReconnectManager` to clarify its intended use.

feat: enhance registry listener signal handling

- Modified the `spawn` function in `registry_listener.rs` to improve signal handling for termination requests, ensuring proper shutdown propagation.

feat: add pooled blocking task variants

- Introduced pooled variants of `spawn_blocking` and `spawn_blocking_value` in `rt/mod.rs` to utilize a shared thread pool for blocking tasks, improving resource management.

fix: ensure stable layout for RpcHeader

- Added tests to verify the layout and maximum name length acceptance of `RpcHeader`, ensuring compatibility with the wire protocol.

test: implement signal shutdown integration test

- Created a new integration test in `signal_shutdown.rs` to verify that the process shuts down cleanly upon receiving a SIGINT signal, utilizing iceoryx2's native signal handling.

+++ ice-rpc/examples/node_liveness_probe.rs
@@ -1,6 +1,6 @@
 //! Diagnostic harness for audit **C7**: validate that iceoryx2's *native* node
-//! monitoring detects a provider killed with `SIGKILL`, before removing the
-//! hand-written kernel lock (`node_lock.rs`).
+//! monitoring detects a provider killed with `SIGKILL`. This is the mechanism
+//! ice-rpc now relies on; the hand-written kernel lock it replaced is gone.
 //!
 //! It is deliberately standalone: it uses the same iceoryx2 service as ice-rpc
 //! (`ipc_threadsafe::Service`) and `Node::list` / `NodeState`.
@@ -19,6 +19,7 @@
 //!
 //! See `scripts/validate-node-liveness.sh` for the orchestrated T1/T2/T3 runs.
 
+#![allow(clippy::unwrap_used)] // tests/examples/benches may panic; production libs keep the deny, see [workspace.lints]
 use std::io::Write;
 use std::time::{Duration, Instant};
 
@@ -43,7 +44,7 @@ fn label<S: Service>(state: &NodeState<S>) -> &'static str {
 fn scan() -> Vec<(u32, &'static str)> {
     let mut out = Vec::new();
     Node::<ipc_threadsafe::Service>::list(Config::global_config(), |state| {
-        out.push((state.node_id().pid().value() as u32, label(&state)));
+        out.push((state.node_id().pid().value(), label(&state)));
         CallbackProgression::Continue
     })
     .expect("Node::list failed");
